Separate pages into their own components for react-router
