+++ web/meetup-facilitator/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@react-oauth/google": "^0.11.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.17.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
