+++ web/meetup-facilitator/src/pages/Login.jsx
@@ -1,0 +1,16 @@
+import GoogleAuth from "../features/GoogleAuth";
+
+function Login() {
+
+
+
+    return (
+        <>
+            <h1>Welcome to Optimeet</h1>
+        <   GoogleAuth/>
+        </>
+        
+    )
+}
+
+export default Login;
